Add scripts for csv to HTTP POST requests
